+++ account-kit/signer/src/base.ts
@@ -21,7 +21,14 @@ import type { Mutate, StoreApi } from "zustand";
 import { subscribeWithSelector } from "zustand/middleware";
 import { createStore } from "zustand/vanilla";
 import type { BaseSignerClient } from "./client/base";
-import type { OauthConfig, OauthParams, User } from "./client/types";
+import type {
+  EmailType,
+  MfaFactor,
+  OauthConfig,
+  OauthParams,
+  User,
+  VerifyMfaParams,
+} from "./client/types";
 import { NotAuthenticatedError } from "./errors.js";
 import { SignerLogger } from "./metrics.js";
 import {
@@ -51,6 +58,10 @@ type AlchemySignerStore = {
   error: ErrorInfo | null;
   otpId?: string;
   isNewUser?: boolean;
+  mfaStatus: {
+    mfaRequired: boolean;
+    mfaFactorId?: string;
+  };
 };
 
 type UnpackedSignature = {
@@ -99,6 +110,10 @@ export abstract class BaseAlchemySigner<TClient extends BaseSignerClient>
             user: null,
             status: AlchemySignerStatus.INITIALIZING,
             error: initialError ?? null,
+            mfaStatus: {
+              mfaRequired: false,
+              mfaFactorId: undefined,
+            },
           } satisfies AlchemySignerStore)
       )
     );
@@ -169,6 +184,13 @@ export abstract class BaseAlchemySigner<TClient extends BaseSignerClient>
           (isNewUser) => {
             if (isNewUser) (listener as AlchemySignerEvents["newUserSignup"])();
           },
+          { fireImmediately: true }
+        );
+      case "mfaStatusChanged":
+        return this.store.subscribe(
+          ({ mfaStatus }) => mfaStatus,
+          (mfaStatus) =>
+            (listener as AlchemySignerEvents["mfaStatusChanged"])(mfaStatus),
           { fireImmediately: true }
         );
       default:
@@ -580,6 +602,39 @@ export abstract class BaseAlchemySigner<TClient extends BaseSignerClient>
     }
   );
 
+  /**
+   * Gets the current MFA status
+   *
+   * @example
+   * ```ts
+   * import { AlchemyWebSigner } from "@account-kit/signer";
+   *
+   * const signer = new AlchemyWebSigner({
+   *  client: {
+   *    connection: {
+   *      rpcUrl: "/api/rpc",
+   *    },
+   *    iframeConfig: {
+   *      iframeContainerId: "alchemy-signer-iframe-container",
+   *    },
+   *  },
+   * });
+   *
+   * const mfaStatus = signer.getMfaStatus();
+   * if (mfaStatus === AlchemyMfaStatus.REQUIRED) {
+   *   // Handle MFA requirement
+   * }
+   * ```
+   *
+   * @returns {{ mfaRequired: boolean; mfaFactorId?: string }} The current MFA status
+   */
+  getMfaStatus = (): {
+    mfaRequired: boolean;
+    mfaFactorId?: string;
+  } => {
+    return this.store.getState().mfaStatus;
+  };
+
   private unpackSignRawMessageBytes = (
     hex: `0x${string}`
   ): UnpackedSignature => {
@@ -769,70 +824,48 @@ export abstract class BaseAlchemySigner<TClient extends BaseSignerClient>
   private authenticateWithEmail = async (
     params: Extract<AuthParams, { type: "email" }>
   ): Promise<User> => {
-    if ("email" in params) {
-      const existingUser = await this.getUser(params.email);
-      const expirationSeconds = this.getExpirationSeconds();
-
-      const { orgId, otpId } = existingUser
-        ? await this.inner.initEmailAuth({
-            email: params.email,
-            emailMode: params.emailMode,
-            expirationSeconds,
-            redirectParams: params.redirectParams,
-          })
-        : await this.inner.createAccount({
-            type: "email",
-            email: params.email,
-            emailMode: params.emailMode,
-            expirationSeconds,
-            redirectParams: params.redirectParams,
-          });
-
-      this.sessionManager.setTemporarySession({
-        orgId,
-        isNewUser: !existingUser,
-      });
-      this.store.setState({
-        status: AlchemySignerStatus.AWAITING_EMAIL_AUTH,
-        otpId,
-        error: null,
-      });
-
-      // We wait for the session manager to emit a connected event if
-      // cross tab sessions are permitted
-      return new Promise<User>((resolve) => {
-        const removeListener = this.sessionManager.on(
-          "connected",
-          (session) => {
-            resolve(session.user);
-            removeListener();
-          }
-        );
-      });
-    } else {
-      const temporarySession = params.orgId
-        ? { orgId: params.orgId }
-        : this.sessionManager.getTemporarySession();
-
-      if (!temporarySession) {
-        this.store.setState({
-          status: AlchemySignerStatus.DISCONNECTED,
-        });
-        throw new Error("Could not find email auth init session!");
-      }
-
-      const user = await this.inner.completeAuthWithBundle({
-        bundle: params.bundle,
-        orgId: temporarySession.orgId,
-        connectedEventName: "connectedEmail",
-        authenticatingType: "email",
-      });
-
-      // fire new user event
-      this.emitNewUserEvent(params.isNewUser);
-
-      return user;
+    if ("bundle" in params) {
+      return this.completeEmailAuth(params);
     }
+
+    if (!("email" in params)) {
+      throw new Error("Email is required");
+    }
+
+    const { orgId, otpId, multiFactors, isNewUser } =
+      await this.initOrCreateEmailUser(
+        params.email,
+        params.emailMode ?? "otp",
+        params.multiFactors,
+        params.redirectParams
+      );
+
+    const isMfaRequired = multiFactors ? multiFactors?.length > 0 : false;
+
+    this.sessionManager.setTemporarySession({
+      orgId,
+      isNewUser,
+      isMfaRequired,
+    });
+
+    this.store.setState({
+      status: AlchemySignerStatus.AWAITING_EMAIL_AUTH,
+      otpId,
+      error: null,
+      mfaStatus: {
+        mfaRequired: isMfaRequired,
+        mfaFactorId: multiFactors?.[0]?.multiFactorId,
+      },
+    });
+
+    // We wait for the session manager to emit a connected event if
+    // cross tab sessions are permitted
+    return new Promise<User>((resolve) => {
+      const removeListener = this.sessionManager.on("connected", (session) => {
+        resolve(session.user);
+        removeListener();
+      });
+    });
   };
 
   private authenticateWithPasskey = async (
@@ -893,7 +926,7 @@ export abstract class BaseAlchemySigner<TClient extends BaseSignerClient>
     args: Extract<AuthParams, { type: "otp" }>
   ): Promise<User> => {
     const tempSession = this.sessionManager.getTemporarySession();
-    const { orgId, isNewUser } = tempSession ?? {};
+    const { orgId, isNewUser, isMfaRequired } = tempSession ?? {};
     const { otpId } = this.store.getState();
     if (!orgId) {
       throw new Error("orgId not found in session");
@@ -901,11 +934,16 @@ export abstract class BaseAlchemySigner<TClient extends BaseSignerClient>
     if (!otpId) {
       throw new Error("otpId not found in session");
     }
+    if (isMfaRequired && !args.multiFactors) {
+      throw new Error(`MFA is required.`);
+    }
+
     const { bundle } = await this.inner.submitOtpCode({
       orgId,
       otpId,
       otpCode: args.otpCode,
       expirationSeconds: this.getExpirationSeconds(),
+      multiFactors: args.multiFactors,
     });
     const user = await this.inner.completeAuthWithBundle({
       bundle,
@@ -1011,6 +1049,79 @@ export abstract class BaseAlchemySigner<TClient extends BaseSignerClient>
     // assumes that if isNewUser is undefined it is a returning user
     if (isNewUser) this.store.setState({ isNewUser });
   };
+
+  private async initOrCreateEmailUser(
+    email: string,
+    emailMode: EmailType,
+    multiFactors?: VerifyMfaParams[],
+    redirectParams?: URLSearchParams
+  ): Promise<{
+    orgId: string;
+    otpId?: string;
+    multiFactors?: MfaFactor[];
+    isNewUser: boolean;
+  }> {
+    const existingUser = await this.getUser(email);
+    const expirationSeconds = this.getExpirationSeconds();
+
+    if (existingUser) {
+      const {
+        orgId,
+        otpId,
+        multiFactors: mfaFactors,
+      } = await this.inner.initEmailAuth({
+        email: email,
+        emailMode: emailMode,
+        expirationSeconds,
+        redirectParams: redirectParams,
+        multiFactors,
+      });
+      return {
+        orgId,
+        otpId,
+        multiFactors: mfaFactors,
+        isNewUser: false,
+      };
+    }
+
+    const { orgId, otpId } = await this.inner.createAccount({
+      type: "email",
+      email,
+      emailMode,
+      expirationSeconds,
+      redirectParams,
+    });
+    return {
+      orgId,
+      otpId,
+      isNewUser: true,
+    };
+  }
+
+  private async completeEmailAuth(
+    params: Extract<AuthParams, { type: "email"; bundle: string }>
+  ): Promise<User> {
+    const temporarySession = params.orgId
+      ? { orgId: params.orgId }
+      : this.sessionManager.getTemporarySession();
+
+    if (!temporarySession) {
+      this.store.setState({ status: AlchemySignerStatus.DISCONNECTED });
+      throw new Error("Could not find email auth init session!");
+    }
+
+    const user = await this.inner.completeAuthWithBundle({
+      bundle: params.bundle,
+      orgId: temporarySession.orgId,
+      connectedEventName: "connectedEmail",
+      authenticatingType: "email",
+    });
+
+    // fire new user event
+    this.emitNewUserEvent(params.isNewUser);
+
+    return user;
+  }
 }
 
 function toErrorInfo(error: unknown): ErrorInfo {