+++ account-kit/signer/src/index.ts
@@ -6,7 +6,11 @@ export {
   OauthFailedError,
 } from "./client/index.js";
 export type * from "./client/types.js";
-export { NotAuthenticatedError, OAuthProvidersError } from "./errors.js";
+export {
+  NotAuthenticatedError,
+  OAuthProvidersError,
+  MfaRequiredError,
+} from "./errors.js";
 export {
   DEFAULT_SESSION_MS,
   SessionManagerParamsSchema,