+++ examples/ui-demo/src/components/small-cards/Wrapper.tsx
@@ -5,23 +5,26 @@ import {
   TransactionsCard7702,
   TransactionsCardDefault,
 } from "./TransactionsCard";
+import { MFACard } from "./MFACard";
 
 export const SmallCardsWrapper = () => {
   const { accountMode } = useConfigStore();
 
   return (
-    <div className="flex flex-col xl:flex-row gap-6 lg:mt-6 items-center p-6 w-full justify-center max-w-screen-sm xl:max-w-none">
+    <div className="flex flex-col xl:flex-row flex-wrap gap-6 lg:mt-6 items-center p-6 w-full justify-center max-w-screen-sm xl:max-w-none">
       {accountMode === "default" ? (
         <>
           <MintCardDefault />
           <SolanaCard />
           <TransactionsCardDefault />
+          <MFACard />
         </>
       ) : (
         <>
           <MintCard7702 />
           <SolanaCard />
           <TransactionsCard7702 />
+          <MFACard />
         </>
       )}
     </div>