+++ site/sidebar/reference/account-kit/react-native.ts
@@ -18,6 +18,10 @@ export const accountKitReactNativeReferenceSidebar: SidebarItem[] = [
         text: "RNAlchemySigner",
         link: "/reference/account-kit/react-native/functions/RNAlchemySigner",
       },
+      {
+        text: "createConfig",
+        link: "/reference/account-kit/react-native/functions/createConfig",
+      },
     ],
   },
   {