+++ account-kit/react-native/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.16.1-alpha.3](https://github.com/alchemyplatform/aa-sdk/compare/v4.16.1-alpha.2...v4.16.1-alpha.3) (2025-03-14)
+
+**Note:** Version bump only for package @account-kit/react-native
+
+## [4.16.1-alpha.2](https://github.com/alchemyplatform/aa-sdk/compare/v4.16.1-alpha.1...v4.16.1-alpha.2) (2025-03-14)
+
+**Note:** Version bump only for package @account-kit/react-native
+
+## [4.16.1-alpha.1](https://github.com/alchemyplatform/aa-sdk/compare/v4.16.0...v4.16.1-alpha.1) (2025-03-14)
+
+### Features
+
+- add RN hooks docs and guides ([#1437](https://github.com/alchemyplatform/aa-sdk/issues/1437)) ([b749c1a](https://github.com/alchemyplatform/aa-sdk/commit/b749c1a7e9a7978aa0327d408a9c7cc724b8cd0f))
+
+## [4.16.1-alpha.0](https://github.com/alchemyplatform/aa-sdk/compare/v4.16.0...v4.16.1-alpha.0) (2025-03-14)
+
+### Features
+
+- add RN hooks docs and guides ([#1437](https://github.com/alchemyplatform/aa-sdk/issues/1437)) ([b749c1a](https://github.com/alchemyplatform/aa-sdk/commit/b749c1a7e9a7978aa0327d408a9c7cc724b8cd0f))
+
 # [4.16.0](https://github.com/alchemyplatform/aa-sdk/compare/v4.15.3...v4.16.0) (2025-03-12)
 
 **Note:** Version bump only for package @account-kit/react-native